Changes done

+++ GarageHub.Domain/Entities/SalesInvoiceItem.cs
@@ -6,7 +6,7 @@ public class SalesInvoiceItem
     public int SaleId { get; set; }
     public int PartId { get; set; }
     public int Quantity { get; set; }
-    public float UnitPrice { get; set; }
+    public decimal UnitPrice { get; set; }
 
     public SalesInvoice SalesInvoice { get; set; } = null!;
 }
